feat(datatable): refactor data table, north arrow component for custom hooks  #1518
